+++ src/js/calculator/types/Gloves/BauerGlovesYT.js
@@ -14,7 +14,7 @@ export function BauerGlovesYT() {
     },
 
     getResult() {
-      const length = this.invert(parseInt(this.input.length, 10));
+      const length = this.invert(parseFloat(this.input.length, 10));
 
       if (length >= 10.5 && length <= 13) {
         return '8';
